python_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pandas==2.2.0
numpy==1.26.4
pandas-ta==0.3.14b0
alpaca-py==0.21.0
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
httpx==0.26.0
matplotlib==3.8.2
plotly==5.18.0
